Fixed a lot
Couldn’t get the pointers with the queue to work; ultimately just
changed it to an enum defining the different users and a switch case to
swap out machine/user registers. Also fixed a little formatting and
commenting here and there.

Undefined DEBUG_VERBOSE for now.

Fixed some formatting with the machine debug output.