Update loop in qtos()
Queues aren't randomly accessible like arrays; they need to be popped and read. Thus we take a backup as we pop through it, and then restore from backup.

Also, and I hadn't caught this the first time around, q.size() will decrease dynamically with the loop; thus the iterator would be increasing and the upper bound would be decreasing, and the loop would stop halfway through the queue. It has now been updated to copy the size into an int.