Implement OS loader
This does not have anything to do with the load function of the
filesystem; this is the OS loader, which requests that the filesystem
retrieve a file and sends it to a buffer. From that buffer, the loader
transfers the file into its correct memory location, before returning
the PCB of the newly loaded process.